+++ tests/preprocessing/test_preprocessors.py
@@ -13,7 +13,7 @@ from dtaianomaly import preprocessing, utils
     preprocessing.MovingAverage(window_size=15),
     preprocessing.NbSamplesUnderSampler(nb_samples=150),
     preprocessing.SamplingRateUnderSampler(sampling_rate=5),
-    preprocessing.ZNormalizer(),
+    preprocessing.StandardScaler(),
     preprocessing.Differencing(order=1),
     preprocessing.PiecewiseAggregateApproximation(n=32)
 ])